% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{add_common_name}
\alias{add_common_name}
\title{Returns dataframe with and additional column (common_name) corresponding
to each animal in the id column of the provided dataframe.}
\usage{
add_common_name(conn, id_df)
}
\arguments{
\item{conn}{database connection}

\item{id_df}{dataframe with animal Ids in id column}
}
\description{
Returns NA in vector if not available.
}
\details{
Assumes access to master database table in animal database.
}
