% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_departure_date.R
\name{add_departure_date}
\alias{add_departure_date}
\title{Returns dataframe with POSIXct date vector of departure dates for each animal
in character vector id of the provided dataframe id_df.
Returns NA in vector if not available.}
\usage{
add_departure_date(conn, id_df)
}
\arguments{
\item{conn}{database connection}

\item{id_df}{dataframe with animal Ids in id column}
}
\description{
Assumes access to master database table in animal database.
}
