% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{add_sex}
\alias{add_sex}
\title{Returns dataframe with sex for each animal
in character vector id of the provided dataframe id_df.
Returns NA in vector if not available.}
\usage{
add_sex(conn, id_df)
}
\arguments{
\item{conn}{database connection}

\item{id_df}{dataframe with animal Ids in id column}
}
\description{
Assumes access to master database table in animal database.
}
