% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_id_to_snprc_id.R
\name{alias_id_to_snprc_id}
\alias{alias_id_to_snprc_id}
\title{Converts an alias (alternate) animal Id to an animal Id.}
\usage{
alias_id_to_snprc_id(conn, id_value)
}
\arguments{
\item{conn}{database connection object}

\item{id_value}{character vector of one or more alternate animal Ids.
Primary Ids are also allowed.}
}
\description{
A dataframe is returned with primary Id (snprc_id), alias Id (id_value),
and alternate Id type (id_type).
}
