% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age_location_sex_species_view.R
\name{create_age_location_sex_species_view}
\alias{create_age_location_sex_species_view}
\title{Creates v_animal_age_location_sex_species as a view in animal database if
it does not already exists.}
\usage{
create_age_location_sex_species_view(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Creates v_animal_age_location_sex_species as a view in animal database if
it does not already exists.
}
