% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{create_Xhalfaversary}
\alias{create_Xhalfaversary}
\title{Create the database table \code{animal.dbl.Xhalfaversary}}
\usage{
create_Xhalfaversary(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Create the database table \code{animal.dbl.Xhalfaversary}
}
