% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_database_record_by_object_id.R
\name{delete_database_record_by_object_id}
\alias{delete_database_record_by_object_id}
\title{Returns logical vector indicating TRUE if record was deleted.}
\usage{
delete_database_record_by_object_id(conn, table, object_id)
}
\arguments{
\item{conn}{database connection object}

\item{table}{character variable having the database table name}

\item{object_id}{integer vector with obr object Id(s) to delete}
}
\description{
This deletes records one at a time, which is expensive in time,
but returns a logical variable indicating success (TRUE) or
failure for each object Id provided.
}
