% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_database_tbl.R
\name{delete_database_tbl}
\alias{delete_database_tbl}
\title{Returns TRUE if the database table was successfully deleted.}
\usage{
delete_database_tbl(conn, table_name, owner = "dbo")
}
\arguments{
\item{conn}{database connnection object}

\item{table_name}{character vector length one having the table name}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns TRUE if the database table was successfully deleted.
}
