% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_id_from_tbl.R
\name{delete_id_from_tbl}
\alias{delete_id_from_tbl}
\title{Returns the number of rows deleted from a database table when all the rows associated
with a list of names are deleted.}
\usage{
delete_id_from_tbl(conn, table_name, ids, owner = "dbo")
}
\arguments{
\item{conn}{database connection object}

\item{table_name}{character vector of length 1 having the table name}

\item{ids}{character vector containing the animal Ids}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns the number of rows deleted from a database table when all the rows associated
with a list of names are deleted.
}
