% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_offspring_from_offspring_tbl.R
\name{delete_offspring_from_offspring_table}
\alias{delete_offspring_from_offspring_table}
\title{Returns the number of rows deleted from the offspring table}
\usage{
delete_offspring_from_offspring_table(conn, ids, owner = "dbo")
}
\arguments{
\item{conn}{database connection object}

\item{ids}{character vector with the animal Ids to be deleted if they are
the offspring_id}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns the number of rows deleted from the offspring table
}
