% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_trigger.R
\name{disable_trigger}
\alias{disable_trigger}
\title{Has the side effect of disabling a specific trigger as defined by type and
table name.}
\usage{
disable_trigger(conn, type, table_name)
}
\arguments{
\item{conn}{database connection object}

\item{type}{one of c("insert", "update", "delete")}

\item{table_name}{character vector with the table name}
}
\description{
Has the side effect of disabling a specific trigger as defined by type and
table name.
}
