% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{drop_Xhalfaversary}
\alias{drop_Xhalfaversary}
\title{Drops Xhalfaversary table and returns TRUE if the database table
\code{animal.dbl.Xhalfaversary} was successfully drop}
\usage{
drop_Xhalfaversary(conn)
}
\arguments{
\item{conn}{database connnection object}
}
\description{
Drops Xhalfaversary table and returns TRUE if the database table
\code{animal.dbl.Xhalfaversary} was successfully drop
}
