% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{fill_halfaversary_tbl}
\alias{fill_halfaversary_tbl}
\title{Fill the database table \code{animal.dbl.Xhalfaversary}}
\usage{
fill_halfaversary_tbl(conn)
}
\arguments{
\item{conn}{database connnection object}
}
\description{
Fill the database table \code{animal.dbl.Xhalfaversary}
}
