% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_age_dist}
\alias{get_age_dist}
\title{Get the maximum age within a species for a specific date range}
\usage{
get_age_dist(conn, arc_species_code, start_date, i)
}
\arguments{
\item{conn}{database connection object}

\item{arc_species_code}{optional character vector of length 1 having the two
character arc_species_code defined by the IACUC. Defaults to "PC".}

\item{start_date}{POSIXct vector of length 1 having the first date
to be included in the time frame of interest.
Should be formatted as "\%m-\%d-\%Y".}

\item{i}{integer used as the offset for setting the base_date being used
in the database query.}
}
\description{
Get the maximum age within a species for a specific date range
}
