% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_assignments_to_iacuc}
\alias{get_assignments_to_iacuc}
\title{Returns dataframe with the animals with a status of assigned, hold, or
 to an IACUC on a particular
date.}
\usage{
get_assignments_to_iacuc(conn, working_iacuc, check_date)
}
\arguments{
\item{conn}{database connection object}

\item{working_iacuc}{IACUC protocol}

\item{check_date}{character string representation of date in mm-dd-yyyy}
}
\description{
Returns dataframe with the animals with a status of assigned, hold, or
 to an IACUC on a particular
date.
}
