% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_audit_proc_df}
\alias{get_audit_proc_df}
\title{Returns dataframe with audit table data for audit_procedures_entered
table of animals for the date range specified.}
\usage{
get_audit_proc_df(conn, ranges)
}
\arguments{
\item{conn}{database connection object}

\item{ranges}{has the list of date ranges and
the animal Ids.}
}
\description{
Returns dataframe with audit table data for audit_procedures_entered
table of animals for the date range specified.
}
