% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_number_of_cagemates.R
\name{get_average_number_of_cagemates}
\alias{get_average_number_of_cagemates}
\title{Returns the average number of cagemates each animal has for days
between and including start and end dates.}
\usage{
get_average_number_of_cagemates(conn, ids, start, end)
}
\arguments{
\item{conn}{database connection}

\item{ids}{character vector with an animal Id in each cell}

\item{start}{character string representation of start date in mm-dd-yyyy
format}

\item{end}{character string representation of end date in mm-dd-yyyy
format}
}
\description{
Returns the average number of cagemates each animal has for days
between and including start and end dates.
}
