% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_births_in_iacuc}
\alias{get_births_in_iacuc}
\title{Returns a dataframe containing the id, sex, and birth date of animals born
between to dates inclusive and assigned to a specific IACUC protocol.}
\usage{
get_births_in_iacuc(conn, working_iacuc, start_date, end_date)
}
\arguments{
\item{conn}{database connection object}

\item{working_iacuc}{IACUC protocol}

\item{start_date}{character vector of length 1 in "mm-dd-YYYY" format}

\item{end_date}{character vector of length 1 in "mm-dd-YYYY" format}
}
\description{
Returns a dataframe containing the id, sex, and birth date of animals born
between to dates inclusive and assigned to a specific IACUC protocol.
}
