% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_cagemates}
\alias{get_cagemates}
\title{Returns the cagemates and total number of days together for each animal
between and including start and end dates.}
\usage{
get_cagemates(conn, ids, start, end)
}
\arguments{
\item{conn}{database connection}

\item{ids}{character vector with an animal Id in each cell}

\item{start}{character string representation of start date in mm-dd-yyyy
format}

\item{end}{character string representation of end date in mm-dd-yyyy
format}
}
\description{
Returns the cagemates and total number of days together for each animal
between and including start and end dates.
}
