% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_charge_account_table}
\alias{get_charge_account_table}
\title{Get charge_account table contents}
\usage{
get_charge_account_table(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Get charge_account table contents
}
