% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_column_names}
\alias{get_column_names}
\title{Returns a character vector containing all of the column names of for the
specified table.}
\usage{
get_column_names(conn, table_name, owner = "DBO")
}
\arguments{
\item{conn}{database connection}

\item{table_name}{character vector with the name of the table}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns a character vector containing all of the column names of for the
specified table.
}
