% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_common_name_for_arc_species_code.R
\name{get_common_name_for_arc_species_codes}
\alias{get_common_name_for_arc_species_codes}
\title{Get common name for arc_species_codes}
\usage{
get_common_name_for_arc_species_codes(conn, arc_species_codes)
}
\arguments{
\item{conn}{database connection object to the animal database}

\item{arc_species_codes}{a character vector with one or more two character
species codes as defined by the IACUC.}
}
\description{
Get common name for arc_species_codes
}
