% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_common_name_for_genus.R
\name{get_common_name_for_genus}
\alias{get_common_name_for_genus}
\title{Make a named vector to translate three letter genus and species codes to
common_names.}
\usage{
get_common_name_for_genus(conn)
}
\arguments{
\item{conn}{database connection object to the animal database}
}
\description{
Makes a named vector with three letter \code{species_code} as the names and
\code{common_name} from the \code{arc_valid_species_codes} database table
are the vector values.
}
\details{
The data are constructed so that the common name is the one associated with
the two character \code{arc_species_code} in the
\code{arc_valid_species_codes} database table.
}
