% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_common_name_for_species_code.R
\name{get_common_name_for_species_code}
\alias{get_common_name_for_species_code}
\title{Returns a character vector of common names using a vector of
three character species codes (species_codes)}
\usage{
get_common_name_for_species_code(conn, species_codes)
}
\arguments{
\item{conn}{database connection object to the animal database}

\item{species_codes}{three character species codes (species_codes)}
}
\description{
Returns a character vector of common names using a vector of
three character species codes (species_codes)
}
