% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dams_with_infants.R
\name{get_dams_with_infants}
\alias{get_dams_with_infants}
\title{Dams with infants within a specified date range of provided.}
\usage{
get_dams_with_infants(conn, arc_species_code,
  current_date = strftime(Sys.Date(), format = "\%m-\%d-\%Y"),
  days_as_infant = 180)
}
\arguments{
\item{conn}{database connection object}

\item{arc_species_code}{character vector of length 1 having two character
species code}

\item{current_date}{optional character vector of length 1 having a valid date
represented by a unambiguous string value that specifies the
date for which the report is current. While most of the time current_date
will indeed be the current date. This allows the user to look at the infants
at a prior date.}

\item{days_as_infant}{number of days that an animal is considered to be
an infant for this purpose.}
}
\description{
Dams with infants within a specified date range of provided.
}
