% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_direct_ancestors.R
\name{get_direct_ancestors}
\alias{get_direct_ancestors}
\title{Get direct ancestors for supplied IDs.}
\usage{
get_direct_ancestors(conn, ids)
}
\arguments{
\item{conn}{database connection object}

\item{ids}{character vector of individual ids}
}
\value{
A dataframe with all direct ancestors for Ids provided.

Takes a database connection and a
character vector of animal Ids and retrieves the direct ancestors for each
id. All returned rows represent a unique individual. The SQL gets animals
until there are no new parents. That is the sires and dams are NULL for
the progenitors.
gets immediate parents for one individual :
--SELECT * FROM dbo.f_get_pedigree(' 30145')
}
\description{
Get direct ancestors for supplied IDs.
}
