% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_dispositions_for_iacuc}
\alias{get_dispositions_for_iacuc}
\title{Returns dataframe with the disposition code, disposition description, and
count of animals with the disposition code within start_date and end_date
inclusive for a specific working_iacuc.}
\usage{
get_dispositions_for_iacuc(conn, working_iacuc, start_date, end_date)
}
\arguments{
\item{conn}{database connection object}

\item{working_iacuc}{IACUC protocol}

\item{start_date}{character vector of length 1 in "mm-dd-YYYY" format}

\item{end_date}{character vector of length 1 in "mm-dd-YYYY" format}
}
\description{
Returns dataframe with the disposition code, disposition description, and
count of animals with the disposition code within start_date and end_date
inclusive for a specific working_iacuc.
}
