% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_female_exposed_to_male_days}
\alias{get_female_exposed_to_male_days}
\title{Get number of days females were exposed to fertile males.}
\usage{
get_female_exposed_to_male_days(
  conn,
  arc_species_code = "PC",
  min_date = "01-01-2014",
  max_date = strftime(Sys.Date(), format = "\%m-\%d-\%Y"),
  male_age = 48,
  female_age = 42,
  gestation_length = 183
)
}
\arguments{
\item{conn}{database connection object}

\item{arc_species_code}{optional character vector of length 1 having the two
character arc_species_code defined by the IACUC. Defaults to "PC".}

\item{min_date}{optional character vector of length 1 having the first date to be
included in the time frame of interest. Defaults to "01-01-2014".
Should be formatted as "\%m-\%d-\%Y".}

\item{max_date}{optional character vector of length 1 having the last date to be
included in the time frame of interest. Defaults to current date.
Should be formatted as "\%m-\%d-\%Y".}

\item{male_age}{integer value in months that indicates the minimum age the
male must be on each day with a female.}

\item{female_age}{integer value in months that indicates the minumum age
the female must be on each day with a male.}

\item{gestation_length}{integer value indicating the gestation length in
days for the species. Animals exposed to the male less than the gestational
length are not included. This can be set to 0 to include all females exposed
to fertile males for any length of time. Defaults to 183.}
}
\value{
dataframe with \code{id} (female Id),
\code{days} (number of days exposed),
\code{current_location}, \code{min_date} (first date of exposure),
and \code{max_date} (last date of exposure).

Males having an attribute of "vasectomized" are not included.
}
\description{
Get number of days females were exposed to fertile males.
}
