% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_id_clinical_admit}
\alias{get_id_clinical_admit}
\title{Get clinical admissions active for an animal on a specified date.}
\usage{
get_id_clinical_admit(conn, id, ob_date)
}
\arguments{
\item{conn}{database connection object}

\item{id}{character vector with animal Ids}

\item{ob_date}{date object}
}
\description{
Get clinical admissions active for an animal on a specified date.
}
