% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_str.R
\name{get_id_str}
\alias{get_id_str}
\title{Returns the character string corresponding to the animal Id columns for
specific tables.}
\usage{
get_id_str(conn, table_names, owner = "dbo")
}
\arguments{
\item{conn}{database connection object}

\item{table_names}{character vector with table names}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns the character string corresponding to the animal Id columns for
specific tables.
}
