% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids_having_specific_character.R
\name{get_ids_having_specific_char}
\alias{get_ids_having_specific_char}
\title{Returns dataframe with animals having a specific character in their
permanent Id.}
\usage{
get_ids_having_specific_char(conn, char, exclude = character(0),
  include = character(0))
}
\arguments{
\item{conn}{database connection object}

\item{char}{character to be matched}

\item{exclude}{character vector having two character species codes of
species to exclude from result set.}

\item{include}{character vector having two character species codes of
species to include in the result set.}
}
\description{
Returns dataframe with animals having a specific character in their
permanent Id.
}
