% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_maintenance_iacuc}
\alias{get_maintenance_iacuc}
\title{Get dataframe with maintenance IACUC protocols including
\code{arc_num_seq}, \code{arc_num_genus}, \code{working_iacuc},
and \code{iacuc_type}.}
\usage{
get_maintenance_iacuc(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Several IACUCs are not research protocols and are designated as maintenance
protocols.
}
