% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_max_age}
\alias{get_max_age}
\title{Get the maximum age within a species for a specific date range}
\usage{
get_max_age(conn, arc_species_code, start_date,
  end_date = strftime(Sys.Date(), format = "\%m-\%d-\%Y"))
}
\arguments{
\item{conn}{database connection object}

\item{arc_species_code}{optional character vector of length 1 having the two
character arc_species_code defined by the IACUC. Defaults to "PC".}

\item{start_date}{character vector of length 1 having the first date
to be included in the time frame of interest.
Should be formatted as "\%m-\%d-\%Y".}

\item{end_date}{optional character vector of length 1 having the last date to be
included in the time frame of interest. Defaults to current date.
Should be formatted as "\%m-\%d-\%Y".}
}
\description{
Get the maximum age within a species for a specific date range
}
