% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonproductive_animals.R
\name{get_non_productive_animals}
\alias{get_non_productive_animals}
\title{Get non-productive animals of a specified sex, arc_species_code, and age.}
\usage{
get_non_productive_animals(conn, sex, arc_species_code, age)
}
\arguments{
\item{conn}{database connection object}

\item{sex}{character vector of length one indicating sex ("M", "F", or
"both").}

\item{arc_species_code}{character vector of length 1 having two character
species code}

\item{age}{numeric value of age indicating that the animals list are to be
at least \code{age} years old.}
}
\description{
Get non-productive animals of a specified sex, arc_species_code, and age.
}
