% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object_ids_in_tbl.R
\name{get_object_ids_in_tbl}
\alias{get_object_ids_in_tbl}
\title{Returns a character vector of OBJECT_ID values from a character vector
(o_ids) that were found to be in the database table being queried.}
\usage{
get_object_ids_in_tbl(conn, o_id, table_name)
}
\arguments{
\item{conn}{database connection object}

\item{o_id}{character vector of object_id values to see if they exist
in the table table_name.}

\item{table_name}{character vector of length one with the table name}
}
\description{
Returns a character vector of OBJECT_ID values from a character vector
(o_ids) that were found to be in the database table being queried.
}
