% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_procs.R
\name{get_proc_df}
\alias{get_proc_df}
\title{Get procedure data for animals within date ranges.}
\usage{
get_proc_df(conn, ranges)
}
\arguments{
\item{conn}{database connection object}

\item{ranges}{has the list of animals with corresponding date ranges.}
}
\value{
A dataframe with admission and procedure data with user name (user_n),
and entry_date_tm of animals for the date range specified.

An example of how to construct the \code{ranges} list follows:

ranges <- list(ids = c(" 12345", "1X1234"),
                dates = list(c("1/2/2014", "12/30/2014"),
                             c("6/07/2010", "6/08/2017")))
}
\description{
Get procedure data for animals within date ranges.
}
