% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rearing.R
\name{get_rearing}
\alias{get_rearing}
\title{Get rearing code and rearing description for Ids provided.}
\usage{
get_rearing(id_df, conn)
}
\arguments{
\item{id_df}{dataframe with a column (\code{id}) made up of a character
vector of individual ids}

\item{conn}{database connection}
}
\description{
Takes a database connection and a
character vector of animal Ids and retrieves the siblings for each
id. All returned rows represent a unique individual. The SQL gets all
siblings of the type specified.
}
