% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_relatives}
\alias{get_relatives}
\title{Returns a data frame containing four columns: id, sire_id,
dam_id, sex when given a character vector of one or more Ids in ids.}
\usage{
get_relatives(conn, ids)
}
\arguments{
\item{conn}{database connection object}

\item{ids}{is the vector of Ids for which their relationships are being
sought.}
}
\description{
Reads the data from the database.
}
