% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_siblings}
\alias{get_siblings}
\title{Returns data frame with specified siblings for Ids provided.}
\usage{
get_siblings(conn, ids, type = "all")
}
\arguments{
\item{conn}{database connection}

\item{ids}{character vector of individual ids}

\item{type}{character element having 'all', 'full', 'half', 'maternal' or
'paternal'.}
}
\description{
Takes a database connection and a
character vector of animal Ids and retrieves the siblings for each
id. All returned rows represent a unique individual. The SQL gets all
siblings of the type specified.
}
