% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_table_names_w_id_col}
\alias{get_table_names_w_id_col}
\title{Returns a character vector of table names for the database connection conn
where the names do not include specified strings, each table has a column
with the name 'id', and the master table is listed first.}
\usage{
get_table_names_w_id_col(conn, expunge_list = character(0), owner = "DBO")
}
\arguments{
\item{conn}{database connection}

\item{expunge_list}{character vector with regexpression patterns to be
expunged (removed) from the list of tables. The expunge_list has no
built in expressions.}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns a character vector of table names for the database connection conn
where the names do not include specified strings, each table has a column
with the name 'id', and the master table is listed first.
}
