% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_table_w_col}
\alias{get_table_w_col}
\title{Returns a character vector of tables names for the database connection conn
where the tables contain a column matching a specific regular expression.}
\usage{
get_table_w_col(conn, pattern, expunge_list = character(0), owner)
}
\arguments{
\item{conn}{database connection object}

\item{pattern}{charcter vector of length 1 having the regular expression
used to match columns.}

\item{expunge_list}{character vector with regexpression patterns to be
expunged (removed) from the list of tables. The expunge_list has no
built in expressions.}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns a character vector of tables names for the database connection conn
where the tables contain a column matching a specific regular expression.
}
