% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_tables_and_columns}
\alias{get_tables_and_columns}
\title{Get tables and columns for a database connection.}
\usage{
get_tables_and_columns(conn, expunge_list = character(0), owner = "DBO")
}
\arguments{
\item{conn}{database connection object}

\item{expunge_list}{character vector with regexpression patterns to be
expunged (removed) from the list of tables. The expunge_list has no
built in expressions.}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Get tables and columns for a database connection.
}
