% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_tables_with_data}
\alias{get_tables_with_data}
\title{Returns a dataframe with a character vector listing the table names and an
integer vector indicating the number of rows for those tables having
at least one record with values for the animals of interest.}
\usage{
get_tables_with_data(conn, potential_tables, ids)
}
\arguments{
\item{conn}{database connection}

\item{potential_tables}{character vector with the table names of tables to
examine for possible data}

\item{ids}{character vector with animal Ids}
}
\description{
Returns a dataframe with a character vector listing the table names and an
integer vector indicating the number of rows for those tables having
at least one record with values for the animals of interest.
}
