% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trigger_name.R
\name{get_trigger_name}
\alias{get_trigger_name}
\title{Returns a character vector with one trigger name af a specific type for each table
name provided.}
\usage{
get_trigger_name(type, table_names)
}
\arguments{
\item{type}{one of c("insert", "update", "delete")}

\item{table_names}{character vector with the table names}
}
\description{
Returns a character vector with one trigger name af a specific type for each table
name provided.
}
