% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_vasectomized_males}
\alias{get_vasectomized_males}
\title{Get vasectomized males of a specified arc_species_code that died after a
specified date.}
\usage{
get_vasectomized_males(conn, arc_species_code,
  died_after_date = strftime(Sys.Date(), format = "\%m-\%d-\%Y"))
}
\arguments{
\item{conn}{database connection object}

\item{arc_species_code}{character vector of length 1 having two character
species code}

\item{died_after_date}{optional character vector of length 1 representing
last date of interest. The male must have died after this date or it is
not returned.}
}
\description{
Get vasectomized males of a specified arc_species_code that died after a
specified date.
}
