% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{get_weight_df}
\alias{get_weight_df}
\title{Returns dataframe with weights, date_tm of collection, user name (user_n),
and entry_date_tm of animals from arrival to death.}
\usage{
get_weight_df(conn, ranges)
}
\arguments{
\item{conn}{database connection object}

\item{ranges}{has the list of date ranges prior to going into the BLS4 and
the animal Ids. The dates are not used for this query.}
}
\description{
Returns dataframe with weights, date_tm of collection, user name (user_n),
and entry_date_tm of animals from arrival to death.
}
