% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{has_id_column}
\alias{has_id_column}
\title{Returns a logical vector indicating whether or not each table contains a
column named 'id'.}
\usage{
has_id_column(conn, table_names, owner = "DBO")
}
\arguments{
\item{conn}{database connection}

\item{table_names}{character vector with the names of tables}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns a logical vector indicating whether or not each table contains a
column named 'id'.
}
