% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{is_animal_id}
\alias{is_animal_id}
\title{Returns logical vector indicating if animal ids are in master.}
\usage{
is_animal_id(conn, ids)
}
\arguments{
\item{conn}{database connection object}

\item{ids}{character vector with ids}
}
\description{
Returns logical vector indicating if animal ids are in master.
}
