% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{is_bad_animal_id}
\alias{is_bad_animal_id}
\title{Returns the proported animal Id(s) that cannot be found in the master table
of the animal database.}
\usage{
is_bad_animal_id(conn, ids)
}
\arguments{
\item{conn}{database connection object}

\item{ids}{character vector with animal Ids}
}
\description{
Returns the proported animal Id(s) that cannot be found in the master table
of the animal database.
}
