% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_database_tbl.R
\name{is_database_tbl}
\alias{is_database_tbl}
\title{Returns TRUE if table exist and FALSE if it does not}
\usage{
is_database_tbl(conn, table_name)
}
\arguments{
\item{conn}{database connection object}

\item{table_name}{character vector of length one having the name of the
table.}
}
\description{
Returns TRUE if table exist and FALSE if it does not
}
