% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_database_view.R
\name{is_database_view}
\alias{is_database_view}
\title{Returns TRUE if view exist and FALSE if it does not}
\usage{
is_database_view(conn, view_name, schema = "DBO")
}
\arguments{
\item{conn}{database connection object}

\item{view_name}{character vector of length one having the name of the
view.}

\item{schema}{character vector of length one having the name of the
schema}
}
\description{
Returns TRUE if view exist and FALSE if it does not
}
