% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{is_location}
\alias{is_location}
\title{Returns logical vector indicating if a location exists.}
\usage{
is_location(conn, locations)
}
\arguments{
\item{conn}{database connection object}

\item{locations}{character vector with ids}
}
\description{
Returns logical vector indicating if a location exists.
}
