% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{is_primate_id}
\alias{is_primate_id}
\title{Returns logical vector indicating if animal ids represent a primate.}
\usage{
is_primate_id(conn, ids)
}
\arguments{
\item{conn}{database connection object}

\item{ids}{character vector with ids}
}
\description{
Returns logical vector indicating if animal ids represent a primate.
}
