% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{remove_housekeeping_columns}
\alias{remove_housekeeping_columns}
\title{Returns a character vector with the column_names vector having had all
housekeeping columns removed.}
\usage{
remove_housekeeping_columns(table_name, column_names)
}
\arguments{
\item{table_name}{name of database table}

\item{column_names}{names of columns within the table}
}
\description{
Returns a character vector with the column_names vector having had all
housekeeping columns removed.
}
