% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{rollback_tran}
\alias{rollback_tran}
\title{Rollsback a transaction, returns FALSE if there is not a
transaction to rollback otherwise TRUE'}
\usage{
rollback_tran(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Rollsback a transaction, returns FALSE if there is not a
transaction to rollback otherwise TRUE'
}
