% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{row_exist}
\alias{row_exist}
\title{Returns logical vector of length one indication whether or not at least
one record exists in the table having an id equal to the value of id.}
\usage{
row_exist(conn, table_name, ids)
}
\arguments{
\item{conn}{database connection}

\item{table_name}{character vector of length one with the name of the table}

\item{ids}{character vector of length one or more with the animal Ids
of interest}
}
\description{
Returns logical vector of length one indication whether or not at least
one record exists in the table having an id equal to the value of id.
}
