% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{rows_with_id_in_table}
\alias{rows_with_id_in_table}
\title{Returns integer vector of length one indicating the number of records that
exists in the table having an id equal to the value of id.}
\usage{
rows_with_id_in_table(conn, table_name, ids)
}
\arguments{
\item{conn}{database connection}

\item{table_name}{character vector of length one with the name of the table}

\item{ids}{character vector of length one or more with the animal Ids
of interest}
}
\description{
Returns integer vector of length one indicating the number of records that
exists in the table having an id equal to the value of id.
}
