% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ids_in_table_by_id.R
\name{update_ids_in_tables_by_id}
\alias{update_ids_in_tables_by_id}
\title{Returns the number of rows updated when old Ids are replaced by replacement Ids
in one or more database tables.}
\usage{
update_ids_in_tables_by_id(conn, table_names, old_ids, replacement_ids,
  owner = "dbo")
}
\arguments{
\item{conn}{database connection object}

\item{table_names}{character vector with one or more table names}

\item{old_ids}{character vector containing the animal Ids to be replaced}

\item{replacement_ids}{character vector containing the replacement animal Ids}

\item{owner}{character vector of length 1 having the database schema owner.
Defaults to "DBO".}
}
\description{
Returns the number of rows updated when old Ids are replaced by replacement Ids
in one or more database tables.
}
