% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalr.R
\name{Xhalfaversary_insert}
\alias{Xhalfaversary_insert}
\title{Inserts animal Ids of all primates with each half birth day date from birth
until death or the current date into the database table
\code{animal.dbl.Xhalfaversary}.}
\usage{
Xhalfaversary_insert(conn)
}
\arguments{
\item{conn}{database connnection object}
}
\description{
Inserts animal Ids of all primates with each half birth day date from birth
until death or the current date into the database table
\code{animal.dbl.Xhalfaversary}.
}
